refactor(runtime): seed PR4 operation boundaries
