+++ src/archetype/app/redaction/__init__.py
@@ -1,18 +1,18 @@
 # Copyright 2026 Vangelis Technologies Inc.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Shared pre-durability secret scanning and redaction authority."""
+"""Compatibility imports for the canonical redaction family."""
 
 from archetype.app.redaction.interfaces import iRedactionService
-from archetype.app.redaction.models import (
+from archetype.redaction import (
     RedactedFile,
     RedactedRecord,
     RedactedText,
     RedactionPolicyConfig,
     RedactionReceipt,
+    RedactionService,
     SecretQuarantineError,
 )
-from archetype.app.redaction.service import RedactionService
 
 __all__ = [
     "RedactedFile",
