+++ src/archetype/episodes/__init__.py
@@ -1,0 +1,16 @@
+# Copyright 2026 Vangelis Technologies Inc.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Episode and trajectory authoring contracts."""
+
+from archetype.episodes.contracts import (
+    ClaudeTranscriptSource,
+    TrajectorySelection,
+    TranscriptIngestionResult,
+)
+
+__all__ = [
+    "ClaudeTranscriptSource",
+    "TrajectorySelection",
+    "TranscriptIngestionResult",
+]
